+++ server/src/parser/newtry/tokenizor/types.ts
@@ -45,10 +45,10 @@ export interface IDiagnosticInfo {
 export type TakeToken = Result<Token, TokenKind.Token>;
 export type TakeDiagnostic = Result<IDiagnosticInfo, TokenKind.Diagnostic>;
 export type TakeComment = Result<Token, TokenKind.Commnet>;
-export type TakeHotkey = Result<Token[], TokenKind.Multi>;
+export type TakeMultiToken = Result<Token[], TokenKind.Multi>;
 
 export type TokenResult = 
 	| TakeToken
 	| TakeDiagnostic
 	| TakeComment
-	| TakeHotkey;
+	| TakeMultiToken;